unify adjusting risk views
