+++ features/earn/aave/manage/services/getManageAaveStateMachine.ts
@@ -83,6 +83,7 @@ export function getManageAaveStateMachine$(
         })
         .withContext({
           token,
+          riskRatio: new RiskRatio(new BigNumber(1), RiskRatio.TYPE.MULITPLE),
           userInput: {
             riskRatio: new RiskRatio(new BigNumber(2), RiskRatio.TYPE.MULITPLE),
             amount: new BigNumber(0),
