+++ features/earn/aave/manage/sidebars/SidebarManageAaveVault.tsx
@@ -20,6 +20,7 @@ import { one, zero } from '../../../../../helpers/zero'
 import { OpenVaultAnimation } from '../../../../../theme/animations'
 import { useManageAaveStateMachineContext } from '../containers/AaveManageStateMachineContext'
 import { ManageAaveEvent, ManageAaveStateMachine, ManageAaveStateMachineState } from '../state'
+import { AdjustRiskView } from '../../common/components/SidebarAdjustRiskView'
 
 export interface ManageAaveVaultProps {
   readonly aaveStateMachine: ManageAaveStateMachine
@@ -138,75 +139,75 @@ function ManageAaveFailureStateView({ state, send }: ManageAaveStateProps) {
   return <SidebarSection {...sidebarSectionProps} />
 }
 
-function ManageAaveEditingStateView({ state, send }: ManageAaveStateProps) {
-  const { t } = useTranslation()
-
-  function convertMultipleToColRatio(multiple: BigNumber): BigNumber {
-    return one.div(multiple.minus(one)).plus(one)
-  }
-
-  function convertColRatioToMultiple(colRatio: BigNumber): BigNumber {
-    return convertMultipleToColRatio(colRatio)
-  }
-
-  const liquidationPriceRatio = one
-  const marketStEthEthPrice = amountFromWei(new BigNumber('968102393798180700'), 'ETH')
-  const minColRatio = new BigNumber(5)
-  const minRisk = convertColRatioToMultiple(minColRatio)
-  const maxRisk = state.context.strategyInfo ? state.context.strategyInfo.maxMultiple : zero
-
-  const sidebarSectionProps: SidebarSectionProps = {
-    title: t('manage-earn.aave.vault-form.title'),
-    content: (
-      <Grid gap={3}>
-        <SliderValuePicker
-          sliderPercentageFill={new BigNumber(0)}
-          leftBoundry={liquidationPriceRatio}
-          leftBoundryFormatter={(value) => value.toFixed(2)}
-          rightBoundry={marketStEthEthPrice}
-          rightBoundryFormatter={(value) => `Current: ${value.toFixed(2)}`}
-          onChange={() => {}}
-          minBoundry={minRisk}
-          maxBoundry={maxRisk}
-          lastValue={new BigNumber(2)}
-          disabled={false}
-          step={0.01}
-          leftLabel={t('open-earn.aave.vault-form.configure-multiple.liquidation-price')}
-        />
-        <Flex
-          sx={{
-            variant: 'text.paragraph4',
-            justifyContent: 'space-between',
-            color: 'neutral80',
-          }}
-        >
-          <Text as="span">{t('open-earn.aave.vault-form.configure-multiple.increase-risk')}</Text>
-          <Text as="span">{t('open-earn.aave.vault-form.configure-multiple.decrease-risk')}</Text>
-        </Flex>
-        <SidebarResetButton clear={() => {}} />
-        <TransactionInformationContainer state={state} send={send} />
-      </Grid>
-    ),
-    primaryButton: {
-      isLoading: false,
-      disabled: true,
-      label: t('manage-earn.aave.vault-form.adjust-risk'),
-      action: () => {
-        send('ADJUST_POSITION')
-      },
-    },
-    textButton: {
-      isLoading: false,
-      disabled: false,
-      label: t('manage-earn.aave.vault-form.close'),
-      action: () => {
-        send('CLOSE_POSITION')
-      },
-    },
-  }
-
-  return <SidebarSection {...sidebarSectionProps} />
-}
+// function ManageAaveEditingStateView({ state, send }: ManageAaveStateProps) {
+//   const { t } = useTranslation()
+//
+//   function convertMultipleToColRatio(multiple: BigNumber): BigNumber {
+//     return one.div(multiple.minus(one)).plus(one)
+//   }
+//
+//   function convertColRatioToMultiple(colRatio: BigNumber): BigNumber {
+//     return convertMultipleToColRatio(colRatio)
+//   }
+//
+//   const liquidationPriceRatio = one
+//   const marketStEthEthPrice = amountFromWei(new BigNumber('968102393798180700'), 'ETH')
+//   const minColRatio = new BigNumber(5)
+//   const minRisk = convertColRatioToMultiple(minColRatio)
+//   const maxRisk = state.context.strategyInfo ? state.context.strategyInfo.maxMultiple : zero
+//
+//   const sidebarSectionProps: SidebarSectionProps = {
+//     title: t('manage-earn.aave.vault-form.title'),
+//     content: (
+//       <Grid gap={3}>
+//         <SliderValuePicker
+//           sliderPercentageFill={new BigNumber(0)}
+//           leftBoundry={liquidationPriceRatio}
+//           leftBoundryFormatter={(value) => value.toFixed(2)}
+//           rightBoundry={marketStEthEthPrice}
+//           rightBoundryFormatter={(value) => `Current: ${value.toFixed(2)}`}
+//           onChange={() => {}}
+//           minBoundry={minRisk}
+//           maxBoundry={maxRisk}
+//           lastValue={new BigNumber(2)}
+//           disabled={false}
+//           step={0.01}
+//           leftLabel={t('open-earn.aave.vault-form.configure-multiple.liquidation-price')}
+//         />
+//         <Flex
+//           sx={{
+//             variant: 'text.paragraph4',
+//             justifyContent: 'space-between',
+//             color: 'neutral80',
+//           }}
+//         >
+//           <Text as="span">{t('open-earn.aave.vault-form.configure-multiple.increase-risk')}</Text>
+//           <Text as="span">{t('open-earn.aave.vault-form.configure-multiple.decrease-risk')}</Text>
+//         </Flex>
+//         <SidebarResetButton clear={() => {}} />
+//         <TransactionInformationContainer state={state} send={send} />
+//       </Grid>
+//     ),
+//     primaryButton: {
+//       isLoading: false,
+//       disabled: true,
+//       label: t('manage-earn.aave.vault-form.adjust-risk'),
+//       action: () => {
+//         send('ADJUST_POSITION')
+//       },
+//     },
+//     textButton: {
+//       isLoading: false,
+//       disabled: false,
+//       label: t('manage-earn.aave.vault-form.close'),
+//       action: () => {
+//         send('CLOSE_POSITION')
+//       },
+//     },
+//   }
+//
+//   return <SidebarSection {...sidebarSectionProps} />
+// }
 
 function ManageAaveSuccessStateView({ state, send }: ManageAaveStateProps) {
   const { t } = useTranslation()
@@ -235,10 +236,32 @@ function ManageAaveSuccessStateView({ state, send }: ManageAaveStateProps) {
 export function SidebarManageAaveVault() {
   const { stateMachine } = useManageAaveStateMachineContext()
   const [state, send] = useActor(stateMachine)
+  const { t } = useTranslation()
 
   switch (true) {
     case state.matches('editing'):
-      return <ManageAaveEditingStateView state={state} send={send} />
+      return (
+        <AdjustRiskView
+          state={state}
+          send={send}
+          primaryButton={{
+            isLoading: false,
+            disabled: true,
+            label: t('manage-earn.aave.vault-form.adjust-risk'),
+            action: () => {
+              send('ADJUST_POSITION')
+            },
+          }}
+          textButton={{
+            isLoading: false,
+            disabled: false,
+            label: t('manage-earn.aave.vault-form.close'),
+            action: () => {
+              send('CLOSE_POSITION')
+            },
+          }}
+        />
+      )
     case state.matches('reviewingClosing'):
       return <ManageAaveReviewingStateView state={state} send={send} />
     case state.matches('reviewingAdjusting'):
