+++ features/earn/aave/open/sidebars/SidebarOpenAaveVault.tsx
@@ -1,29 +1,18 @@
-import { RiskRatio } from '@oasisdex/oasis-actions'
 import { useActor } from '@xstate/react'
-import { BigNumber } from 'bignumber.js'
 import { SidebarSection, SidebarSectionProps } from 'components/sidebar/SidebarSection'
 import { useFeatureToggle } from 'helpers/useFeatureToggle'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
-import { Box, Flex, Grid, Image, Link, Text } from 'theme-ui'
+import { Box, Flex, Grid, Image } from 'theme-ui'
 import { Sender } from 'xstate'
-
-import { SliderValuePicker } from '../../../../../components/dumb/SliderValuePicker'
-import { MessageCard } from '../../../../../components/MessageCard'
-import { SidebarResetButton } from '../../../../../components/vault/sidebar/SidebarResetButton'
-import {
-  getEstimatedGasFeeTextOld,
-  VaultChangesInformationContainer,
-  VaultChangesInformationItem,
-} from '../../../../../components/vault/VaultChangesInformation'
-import { formatPercent } from '../../../../../helpers/formatters/format'
 import { staticFilesRuntimeUrl } from '../../../../../helpers/staticPaths'
-import { one, zero } from '../../../../../helpers/zero'
 import { OpenVaultAnimation } from '../../../../../theme/animations'
 import { ProxyView } from '../../../../proxyNew'
 import { useOpenAaveStateMachineContext } from '../containers/AaveOpenStateMachineContext'
 import { OpenAaveEvent, OpenAaveStateMachine, OpenAaveStateMachineState } from '../state/'
 import { SidebarOpenAaveVaultEditingState } from './SidebarOpenAaveVaultEditingState'
+import { AdjustRiskView } from '../../common/components/SidebarAdjustRiskView'
+import { OpenAaveInformationContainer } from '../../common/components/OpenAaveInformationContainer'
 
 export interface OpenAaveVaultProps {
   readonly aaveStateMachine: OpenAaveStateMachine
@@ -34,19 +23,6 @@ interface OpenAaveStateProps {
   readonly send: Sender<OpenAaveEvent>
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function OpenAaveInformationContainer({ state, send }: OpenAaveStateProps) {
-  const { t } = useTranslation()
-  return (
-    <VaultChangesInformationContainer title="Order information">
-      <VaultChangesInformationItem
-        label={t('transaction-fee')}
-        value={getEstimatedGasFeeTextOld(state.context.estimatedGasPrice)}
-      />
-    </VaultChangesInformationContainer>
-  )
-}
-
 function OpenAaveTransactionInProgressStateView({ state, send }: OpenAaveStateProps) {
   const { t } = useTranslation()
 
@@ -55,7 +31,7 @@ function OpenAaveTransactionInProgressStateView({ state, send }: OpenAaveStatePr
     content: (
       <Grid gap={3}>
         <OpenVaultAnimation />
-        <OpenAaveInformationContainer state={state} send={send} />
+        <OpenAaveInformationContainer state={state} />
       </Grid>
     ),
     primaryButton: {
@@ -76,7 +52,7 @@ function OpenAaveReviewingStateView({ state, send }: OpenAaveStateProps) {
     title: t('open-earn.aave.vault-form.title'),
     content: (
       <Grid gap={3}>
-        <OpenAaveInformationContainer state={state} send={send} />
+        <OpenAaveInformationContainer state={state} />
       </Grid>
     ),
     primaryButton: {
@@ -98,7 +74,7 @@ function OpenAaveFailureStateView({ state, send }: OpenAaveStateProps) {
     title: t('open-earn.aave.vault-form.title'),
     content: (
       <Grid gap={3}>
-        <OpenAaveInformationContainer state={state} send={send} />
+        <OpenAaveInformationContainer state={state} />
       </Grid>
     ),
     primaryButton: {
@@ -124,7 +100,7 @@ function OpenAaveEditingStateView({ state, send }: OpenAaveStateProps) {
     content: (
       <Grid gap={3}>
         <SidebarOpenAaveVaultEditingState state={state} send={send} />
-        <OpenAaveInformationContainer state={state} send={send} />
+        <OpenAaveInformationContainer state={state} />
       </Grid>
     ),
     primaryButton: {
@@ -151,7 +127,7 @@ function OpenAaveSuccessStateView({ state, send }: OpenAaveStateProps) {
             <Image src={staticFilesRuntimeUrl('/static/img/protection_complete_v2.svg')} />
           </Flex>
         </Box>
-        <OpenAaveInformationContainer state={state} send={send} />
+        <OpenAaveInformationContainer state={state} />
       </Grid>
     ),
     primaryButton: {
@@ -163,150 +139,10 @@ function OpenAaveSuccessStateView({ state, send }: OpenAaveStateProps) {
   return <SidebarSection {...sidebarSectionProps} />
 }
 
-function SettingMultipleView({ state, send }: OpenAaveStateProps) {
-  const { t } = useTranslation()
-
-  const maxRisk = state.context.transactionParameters?.simulation.position.category.maxLoanToValue
-
-  const minRisk =
-    state.context.transactionParameters?.simulation.minConfigurableRiskRatio ||
-    new RiskRatio(zero, RiskRatio.TYPE.LTV)
-
-  const liquidationPrice =
-    state.context.transactionParameters?.simulation.position.liquidationPrice || zero
-
-  const oracleAssetPrice = state.context.strategyInfo?.oracleAssetPrice || zero
-
-  enum RiskLevel {
-    OK = 'OK',
-    AT_RISK = 'AT_RISK',
-  }
-
-  const healthFactor = state.context.transactionParameters?.simulation.position.healthFactor
-
-  const warningHealthFactor = new BigNumber('1.25')
-
-  const riskTrafficLight = healthFactor?.gt(warningHealthFactor) ? RiskLevel.OK : RiskLevel.AT_RISK
-
-  const collateralToken = state.context.strategyInfo?.collateralToken
-
-  const debtToken = state.context.token
-
-  const priceMovementUntilLiquidation = one.minus(one.div(healthFactor || zero)).times(100)
-
-  const priceMovementWarningThreshold = new BigNumber(20)
-
-  const priceMovementToDisplay = formatPercent(
-    BigNumber.min(priceMovementUntilLiquidation, priceMovementWarningThreshold),
-    { precision: 2 },
-  )
-
-  const isWarning = priceMovementUntilLiquidation.lte(priceMovementWarningThreshold)
-
-  console.log(
-    `state.context.strategyInfo?.liquidationBonus ${state.context.strategyInfo?.liquidationBonus}`,
-  )
-
-  const liquidationPenalty = formatPercent(
-    (state.context.strategyInfo?.liquidationBonus || zero).times(100),
-    {
-      precision: 2,
-    },
-  )
-
-  const sidebarSectionProps: SidebarSectionProps = {
-    title: t('open-earn.aave.vault-form.title'),
-    content: (
-      <Grid gap={3}>
-        <SliderValuePicker
-          sliderPercentageFill={new BigNumber(0)}
-          leftBoundry={liquidationPrice}
-          leftBoundryFormatter={(value) => value.toFixed(2)}
-          rightBoundry={oracleAssetPrice}
-          rightBoundryFormatter={(value) => `Current: ${value.toFixed(2)}`}
-          rightBoundryStyling={{
-            color: riskTrafficLight === RiskLevel.OK ? 'success100' : 'warning100',
-          }}
-          onChange={(ltv) => {
-            send({ type: 'SET_RISK_RATIO', riskRatio: new RiskRatio(ltv, RiskRatio.TYPE.LTV) })
-          }}
-          minBoundry={minRisk.loanToValue || zero}
-          maxBoundry={maxRisk || zero}
-          lastValue={state.context.riskRatio.loanToValue}
-          disabled={false}
-          step={0.01}
-          leftLabel={t('open-earn.aave.vault-form.configure-multiple.liquidation-price', {
-            collateralToken,
-            debtToken,
-          })}
-          rightLabel={
-            <Link target="_blank" href="https://dune.com/dataalways/stETH-De-Peg">
-              <Text variant="paragraph4" color="interactive100">
-                {t('open-earn.aave.vault-form.configure-multiple.historical-ratio', {
-                  collateralToken,
-                  debtToken,
-                })}{' '}
-                &gt;
-              </Text>
-            </Link>
-          }
-        />
-        <Flex
-          sx={{
-            variant: 'text.paragraph4',
-            justifyContent: 'space-between',
-            color: 'neutral80',
-          }}
-        >
-          <Text as="span">{t('open-earn.aave.vault-form.configure-multiple.increase-risk')}</Text>
-          <Text as="span">{t('open-earn.aave.vault-form.configure-multiple.decrease-risk')}</Text>
-        </Flex>
-        <OpenAaveInformationContainer state={state} send={send} />
-        <MessageCard
-          messages={[
-            isWarning
-              ? t('open-earn.aave.vault-form.configure-multiple.vault-message-warning', {
-                  collateralToken,
-                  priceMovement: priceMovementToDisplay,
-                  debtToken,
-                  liquidationPenalty,
-                })
-              : t('open-earn.aave.vault-form.configure-multiple.vault-message-ok', {
-                  collateralToken,
-                  priceMovement: priceMovementToDisplay,
-                  debtToken,
-                  liquidationPenalty,
-                }),
-          ]}
-          type={isWarning ? 'warning' : 'ok'}
-        />
-        <SidebarResetButton
-          clear={() => {
-            send({ type: 'SET_RISK_RATIO', riskRatio: minRisk })
-          }}
-        />
-        <OpenAaveInformationContainer state={state} send={send} />
-      </Grid>
-    ),
-    primaryButton: {
-      steps: [2, state.context.totalSteps!],
-      isLoading: false,
-      disabled: false,
-      label: t('open-earn.aave.vault-form.open-btn'),
-      action: () => send('NEXT_STEP'),
-    },
-    textButton: {
-      label: 'Back to enter ETH',
-      action: () => send('BACK_TO_EDITING'),
-    },
-  }
-
-  return <SidebarSection {...sidebarSectionProps} />
-}
-
 export function SidebarOpenAaveVault() {
   const { stateMachine } = useOpenAaveStateMachineContext()
   const [state, send] = useActor(stateMachine)
+  const { t } = useTranslation()
 
   switch (true) {
     case state.matches('editing'):
@@ -314,7 +150,23 @@ export function SidebarOpenAaveVault() {
     case state.matches('proxyCreating'):
       return <ProxyView proxyMachine={state.context.refProxyMachine!} />
     case state.matches('settingMultiple'):
-      return <SettingMultipleView state={state} send={send} />
+      return (
+        <AdjustRiskView
+          state={state}
+          send={send}
+          primaryButton={{
+            steps: [2, state.context.totalSteps!],
+            isLoading: false,
+            disabled: false,
+            label: t('open-earn.aave.vault-form.open-btn'),
+            action: () => send('NEXT_STEP'),
+          }}
+          textButton={{
+            label: 'Back to enter ETH',
+            action: () => send('BACK_TO_EDITING'),
+          }}
+        />
+      )
     case state.matches('reviewing'):
       return <OpenAaveReviewingStateView state={state} send={send} />
     case state.matches('txInProgress'):
